+++ components/members/main/MemberDetail/index.tsx
@@ -18,9 +18,9 @@ import EmptyProfile from '@/components/members/detail/EmptyProfile';
 import InfoItem from '@/components/members/detail/InfoItem';
 import MemberProjectCard from '@/components/members/detail/MemberProjectCard';
 import PartItem from '@/components/members/detail/PartItem';
-import InterestSection from '@/components/members/main/MemberDetail/InterestSection';
-import MessageSection from '@/components/members/main/MemberDetail/MessageSection';
 import MemberDetailSection from '@/components/members/main/MemberDetail/Section';
+import InterestSection from '@/components/members/main/MemberDetail/Section/Interest';
+import MessageSection from '@/components/members/main/MemberDetail/Section/Message';
 import { DEFAULT_DATE } from '@/components/members/upload/constants';
 import { playgroundLink } from '@/constants/links';
 import { useRunOnce } from '@/hooks/useRunOnce';
